Remove all implicit conversion classes (closes #108)
