+++ src/main/scala/com.snowplowanalytics.iglu/client/Resolver.scala
@@ -36,7 +36,6 @@ import repositories.{EmbeddedRepositoryRef, HttpRepositoryRef, RepositoryRef}
 import validation.SchemaValidation.{getErrors, isValid}
 import validation.ValidatableCirceMethods
 import validation.ProcessingMessage
-import validation.ProcessingMessageMethods._
 
 /**
  * Companion object. Lets us create a Resolver from
@@ -155,7 +154,7 @@ object Resolver {
     } else if (HttpRepositoryRef.isHttp(rc)) {
       HttpRepositoryRef.parse(rc)
     } else {
-      s"Configuration unrecognizable as either embedded or HTTP repository".invalid.toProcessingMessageNel
+      ProcessingMessage(s"Configuration unrecognizable as either embedded or HTTP repository").invalidNel
     }
   }
