+++ src/main/scala/com.snowplowanalytics.iglu/client/validation/ProcessingMessage.scala
@@ -1,0 +1,34 @@
+/*
+ * Copyright (c) 2014-2018 Snowplow Analytics Ltd. All rights reserved.
+ *
+ * This program is licensed to you under the Apache License Version 2.0,
+ * and you may not use this file except in compliance with the Apache License Version 2.0.
+ * You may obtain a copy of the Apache License Version 2.0 at http://www.apache.org/licenses/LICENSE-2.0.
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the Apache License Version 2.0 is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the Apache License Version 2.0 for the specific language governing permissions and limitations there under.
+ */
+package com.snowplowanalytics.iglu.client
+package validation
+
+// Circe
+import io.circe._
+import io.circe.syntax._
+
+// TODO: replace with better format after discussion
+case class ProcessingMessage(
+  message: String,
+  schemaKey: Option[String] = None,
+  repositories: Option[Json] = None) {
+
+  def asJson: Json = {
+    Json.obj(
+      "level" := "error",
+      "message" := message,
+      "schemaKey" := schemaKey
+    )
+  }
+
+}
